refactor UART driver
