+++ esp-hal/src/uart/mod.rs
@@ -47,16 +47,17 @@ crate::unstable_driver! {
     pub mod uhci;
 }
 
+mod compat;
+mod low_level;
+
 use core::{marker::PhantomData, sync::atomic::Ordering, task::Poll};
 
 use enumset::{EnumSet, EnumSetType};
-use portable_atomic::AtomicBool;
 
 use crate::{
     Async,
     Blocking,
     DriverMode,
-    asynch::AtomicWaker,
     gpio::{
         InputConfig,
         InputSignal,
@@ -74,6 +75,52 @@ use crate::{
     soc::clocks::{self, ClockTree},
     system::PeripheralGuard,
 };
+
+use low_level::{sync_regs, Info, Instance, State, UartClockGuard, UartRxFuture, UartTxFuture};
+use low_level::{rx_event_check_for_error, RxEvent, TxEvent};
+
+crate::any_peripheral! {
+    /// Any UART peripheral.
+    pub peripheral AnyUart<'d> {
+        #[cfg(soc_has_uart0)]
+        Uart0(crate::peripherals::UART0<'d>),
+        #[cfg(soc_has_uart1)]
+        Uart1(crate::peripherals::UART1<'d>),
+        #[cfg(soc_has_uart2)]
+        Uart2(crate::peripherals::UART2<'d>),
+        #[cfg(soc_has_uart3)]
+        Uart3(crate::peripherals::UART3<'d>),
+        #[cfg(soc_has_uart4)]
+        Uart4(crate::peripherals::UART4<'d>),
+    }
+}
+
+impl Instance for AnyUart<'_> {
+    #[inline]
+    fn parts(&self) -> (&'static Info, &'static State) {
+        any::delegate!(self, uart => { uart.parts() })
+    }
+}
+
+impl AnyUart<'_> {
+    pub(super) fn bind_peri_interrupt(&self, handler: InterruptHandler) {
+        any::delegate!(self, uart => { uart.bind_peri_interrupt(handler) })
+    }
+
+    pub(super) fn disable_peri_interrupt_on_all_cores(&self) {
+        any::delegate!(self, uart => { uart.disable_peri_interrupt_on_all_cores() })
+    }
+
+    pub(super) fn set_interrupt_handler(&self, handler: InterruptHandler) {
+        self.disable_peri_interrupt_on_all_cores();
+
+        self.info().enable_listen(EnumSet::all(), false);
+        self.info().clear_interrupts(EnumSet::all());
+
+        self.bind_peri_interrupt(handler);
+    }
+}
+
 
 /// UART RX Error
 #[derive(Debug, Clone, Copy, PartialEq)]
@@ -121,20 +168,6 @@ impl core::fmt::Display for RxError {
     }
 }
 
-#[instability::unstable]
-impl embedded_io_06::Error for RxError {
-    fn kind(&self) -> embedded_io_06::ErrorKind {
-        embedded_io_06::ErrorKind::Other
-    }
-}
-
-#[instability::unstable]
-impl embedded_io_07::Error for RxError {
-    fn kind(&self) -> embedded_io_07::ErrorKind {
-        embedded_io_07::ErrorKind::Other
-    }
-}
-
 /// UART TX Error
 #[derive(Debug, Clone, Copy, PartialEq)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
@@ -148,19 +181,6 @@ impl core::fmt::Display for TxError {
 }
 
 impl core::error::Error for TxError {}
-
-#[instability::unstable]
-impl embedded_io_06::Error for TxError {
-    fn kind(&self) -> embedded_io_06::ErrorKind {
-        embedded_io_06::ErrorKind::Other
-    }
-}
-#[instability::unstable]
-impl embedded_io_07::Error for TxError {
-    fn kind(&self) -> embedded_io_07::ErrorKind {
-        embedded_io_07::ErrorKind::Other
-    }
-}
 
 #[instability::unstable]
 pub use crate::soc::clocks::UartFunctionClockSclk as ClockSource;
@@ -585,46 +605,6 @@ impl core::fmt::Display for ConfigError {
     }
 }
 
-#[instability::unstable]
-impl<Dm> embassy_embedded_hal::SetConfig for Uart<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    type Config = Config;
-    type ConfigError = ConfigError;
-
-    fn set_config(&mut self, config: &Self::Config) -> Result<(), Self::ConfigError> {
-        self.apply_config(config)
-    }
-}
-
-#[instability::unstable]
-impl<Dm> embassy_embedded_hal::SetConfig for UartRx<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    type Config = Config;
-    type ConfigError = ConfigError;
-
-    fn set_config(&mut self, config: &Self::Config) -> Result<(), Self::ConfigError> {
-        self.apply_config(config)
-    }
-}
-
-#[instability::unstable]
-impl<Dm> embassy_embedded_hal::SetConfig for UartTx<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    type Config = Config;
-    type ConfigError = ConfigError;
-
-    fn set_config(&mut self, config: &Self::Config) -> Result<(), Self::ConfigError> {
-        self.baudrate = config.baudrate;
-        self.apply_config(config)
-    }
-}
-
 impl<'d> UartTx<'d, Blocking> {
     #[procmacros::doc_replace]
     /// Create a new UART TX instance in [`Blocking`] mode.
@@ -904,12 +884,7 @@ where
     /// Returns `true` if the transmit line is idle, meaning no data is
     /// currently being transmitted.
     fn is_tx_idle(&self) -> bool {
-        #[cfg(esp32)]
-        let status = self.regs().status();
-        #[cfg(not(esp32))]
-        let status = self.regs().fsm_status();
-
-        status.read().st_utx_out().bits() == 0x0
+        self.uart.info().is_tx_idle()
     }
 
     /// Disables all TX-related interrupts for this UART instance.
@@ -935,26 +910,6 @@ where
 
     fn regs(&self) -> &RegisterBlock {
         self.uart.info().regs()
-    }
-}
-
-#[inline(always)]
-fn sync_regs(_register_block: &RegisterBlock) {
-    #[cfg(not(any(esp32, esp32s2)))]
-    {
-        cfg_if::cfg_if! {
-            if #[cfg(any(esp32c2, esp32c3, esp32s3))] {
-                let update_reg = _register_block.id();
-            } else {
-                let update_reg = _register_block.reg_update();
-            }
-        }
-
-        update_reg.modify(|_, w| w.reg_update().set_bit());
-
-        while update_reg.read().reg_update().bit_is_set() {
-            core::hint::spin_loop();
-        }
     }
 }
 
@@ -1100,14 +1055,7 @@ impl<'d> UartRx<'d, Async> {
                 events |= RxEvent::CmdCharDetected;
             }
 
-            cfg_if::cfg_if! {
-                if #[cfg(uart_version = "2")] {
-                    let reg_en = self.regs().tout_conf();
-                } else if #[cfg(uart_version = "1")] {
-                    let reg_en = self.regs().conf1();
-                }
-            };
-            if listen_for_timeout && reg_en.read().rx_tout_en().bit_is_set() {
+            if listen_for_timeout && self.uart.info().rx_timeout_enabled() {
                 events |= RxEvent::FifoTout;
             }
 
@@ -2043,11 +1991,8 @@ where
     /// Configures the AT-CMD detection settings
     #[instability::unstable]
     pub fn set_at_cmd(&mut self, config: AtCmdConfig) {
-        // P4: sclk_en is controlled by HP_SYS_CLKRST, not UART clk_conf register
-        #[cfg(not(any(esp32, esp32s2, esp32p4)))]
-        self.regs()
-            .clk_conf()
-            .modify(|_, w| w.sclk_en().clear_bit());
+        #[cfg(uart_has_sclk_enable)]
+        self.rx.uart.info().set_at_cmd_clock_enabled(false);
 
         self.regs().at_cmd_char().write(|w| unsafe {
             w.at_cmd_char().bits(config.cmd_char);
@@ -2072,8 +2017,8 @@ where
                 .write(|w| unsafe { w.rx_gap_tout().bits(gap_timeout as _) });
         }
 
-        #[cfg(not(any(esp32, esp32s2, esp32p4)))]
-        self.regs().clk_conf().modify(|_, w| w.sclk_en().set_bit());
+        #[cfg(uart_has_sclk_enable)]
+        self.rx.uart.info().set_at_cmd_clock_enabled(true);
 
         sync_regs(self.regs());
     }
@@ -2109,67 +2054,6 @@ where
     }
 }
 
-impl crate::private::Sealed for Uart<'_, Blocking> {}
-
-#[instability::unstable]
-impl crate::interrupt::InterruptConfigurable for Uart<'_, Blocking> {
-    fn set_interrupt_handler(&mut self, handler: InterruptHandler) {
-        // `self.tx.uart` and `self.rx.uart` are the same
-        self.tx.uart.set_interrupt_handler(handler);
-    }
-}
-
-#[instability::unstable]
-impl<Dm> ufmt_write::uWrite for Uart<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    type Error = TxError;
-
-    #[inline]
-    fn write_str(&mut self, s: &str) -> Result<(), Self::Error> {
-        self.tx.write_str(s)
-    }
-
-    #[inline]
-    fn write_char(&mut self, ch: char) -> Result<(), Self::Error> {
-        self.tx.write_char(ch)
-    }
-}
-
-#[instability::unstable]
-impl<Dm> ufmt_write::uWrite for UartTx<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    type Error = TxError;
-
-    #[inline]
-    fn write_str(&mut self, s: &str) -> Result<(), Self::Error> {
-        self.write_all(s.as_bytes())
-    }
-}
-
-impl<Dm> core::fmt::Write for Uart<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    #[inline]
-    fn write_str(&mut self, s: &str) -> core::fmt::Result {
-        self.tx.write_str(s)
-    }
-}
-
-impl<Dm> core::fmt::Write for UartTx<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    #[inline]
-    fn write_str(&mut self, s: &str) -> core::fmt::Result {
-        self.write_all(s.as_bytes()).map_err(|_| core::fmt::Error)
-    }
-}
-
 /// UART Tx or Rx Error
 #[instability::unstable]
 #[derive(Debug, Clone, Copy, PartialEq)]
@@ -2196,20 +2080,6 @@ impl core::fmt::Display for IoError {
 }
 
 #[instability::unstable]
-impl embedded_io_06::Error for IoError {
-    fn kind(&self) -> embedded_io_06::ErrorKind {
-        embedded_io_06::ErrorKind::Other
-    }
-}
-
-#[instability::unstable]
-impl embedded_io_07::Error for IoError {
-    fn kind(&self) -> embedded_io_07::ErrorKind {
-        embedded_io_07::ErrorKind::Other
-    }
-}
-
-#[instability::unstable]
 impl From<RxError> for IoError {
     fn from(e: RxError) -> Self {
         IoError::Rx(e)
@@ -2220,490 +2090,6 @@ impl From<RxError> for IoError {
 impl From<TxError> for IoError {
     fn from(e: TxError) -> Self {
         IoError::Tx(e)
-    }
-}
-
-#[instability::unstable]
-impl<Dm: DriverMode> embedded_io_06::ErrorType for Uart<'_, Dm> {
-    type Error = IoError;
-}
-
-#[instability::unstable]
-impl<Dm: DriverMode> embedded_io_06::ErrorType for UartTx<'_, Dm> {
-    type Error = TxError;
-}
-
-#[instability::unstable]
-impl<Dm: DriverMode> embedded_io_06::ErrorType for UartRx<'_, Dm> {
-    type Error = RxError;
-}
-
-#[instability::unstable]
-impl<Dm> embedded_io_06::Read for Uart<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    fn read(&mut self, buf: &mut [u8]) -> Result<usize, Self::Error> {
-        self.rx.read(buf).map_err(IoError::Rx)
-    }
-}
-
-#[instability::unstable]
-impl<Dm> embedded_io_06::Read for UartRx<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    fn read(&mut self, buf: &mut [u8]) -> Result<usize, Self::Error> {
-        self.read(buf)
-    }
-}
-
-#[instability::unstable]
-impl<Dm> embedded_io_06::ReadReady for Uart<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    fn read_ready(&mut self) -> Result<bool, Self::Error> {
-        Ok(self.rx.read_ready())
-    }
-}
-
-#[instability::unstable]
-impl<Dm> embedded_io_06::ReadReady for UartRx<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    fn read_ready(&mut self) -> Result<bool, Self::Error> {
-        Ok(UartRx::read_ready(self))
-    }
-}
-
-#[instability::unstable]
-impl<Dm> embedded_io_06::Write for Uart<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    fn write(&mut self, buf: &[u8]) -> Result<usize, Self::Error> {
-        self.tx.write(buf).map_err(IoError::Tx)
-    }
-
-    fn flush(&mut self) -> Result<(), Self::Error> {
-        self.tx.flush().map_err(IoError::Tx)
-    }
-}
-
-#[instability::unstable]
-impl<Dm> embedded_io_06::Write for UartTx<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    fn write(&mut self, buf: &[u8]) -> Result<usize, Self::Error> {
-        self.write(buf)
-    }
-
-    fn flush(&mut self) -> Result<(), Self::Error> {
-        self.flush()
-    }
-}
-
-#[instability::unstable]
-impl<Dm> embedded_io_06::WriteReady for UartTx<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    fn write_ready(&mut self) -> Result<bool, Self::Error> {
-        Ok(UartTx::write_ready(self))
-    }
-}
-
-#[instability::unstable]
-impl<Dm> embedded_io_06::WriteReady for Uart<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    fn write_ready(&mut self) -> Result<bool, Self::Error> {
-        Ok(self.tx.write_ready())
-    }
-}
-
-#[instability::unstable]
-impl<Dm: DriverMode> embedded_io_07::ErrorType for Uart<'_, Dm> {
-    type Error = IoError;
-}
-
-#[instability::unstable]
-impl<Dm: DriverMode> embedded_io_07::ErrorType for UartTx<'_, Dm> {
-    type Error = TxError;
-}
-
-#[instability::unstable]
-impl<Dm: DriverMode> embedded_io_07::ErrorType for UartRx<'_, Dm> {
-    type Error = RxError;
-}
-
-#[instability::unstable]
-impl<Dm> embedded_io_07::Read for Uart<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    fn read(&mut self, buf: &mut [u8]) -> Result<usize, Self::Error> {
-        self.rx.read(buf).map_err(IoError::Rx)
-    }
-}
-
-#[instability::unstable]
-impl<Dm> embedded_io_07::Read for UartRx<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    fn read(&mut self, buf: &mut [u8]) -> Result<usize, Self::Error> {
-        self.read(buf)
-    }
-}
-
-#[instability::unstable]
-impl<Dm> embedded_io_07::ReadReady for Uart<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    fn read_ready(&mut self) -> Result<bool, Self::Error> {
-        Ok(self.rx.read_ready())
-    }
-}
-
-#[instability::unstable]
-impl<Dm> embedded_io_07::ReadReady for UartRx<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    fn read_ready(&mut self) -> Result<bool, Self::Error> {
-        Ok(UartRx::read_ready(self))
-    }
-}
-
-#[instability::unstable]
-impl<Dm> embedded_io_07::Write for Uart<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    fn write(&mut self, buf: &[u8]) -> Result<usize, Self::Error> {
-        self.tx.write(buf).map_err(IoError::Tx)
-    }
-
-    fn flush(&mut self) -> Result<(), Self::Error> {
-        self.tx.flush().map_err(IoError::Tx)
-    }
-}
-
-#[instability::unstable]
-impl<Dm> embedded_io_07::Write for UartTx<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    fn write(&mut self, buf: &[u8]) -> Result<usize, Self::Error> {
-        self.write(buf)
-    }
-
-    fn flush(&mut self) -> Result<(), Self::Error> {
-        self.flush()
-    }
-}
-
-#[instability::unstable]
-impl<Dm> embedded_io_07::WriteReady for UartTx<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    fn write_ready(&mut self) -> Result<bool, Self::Error> {
-        Ok(UartTx::write_ready(self))
-    }
-}
-
-#[instability::unstable]
-impl<Dm> embedded_io_07::WriteReady for Uart<'_, Dm>
-where
-    Dm: DriverMode,
-{
-    fn write_ready(&mut self) -> Result<bool, Self::Error> {
-        Ok(self.tx.write_ready())
-    }
-}
-
-#[derive(Debug, EnumSetType)]
-pub(crate) enum TxEvent {
-    Done,
-    FiFoEmpty,
-}
-
-#[derive(Debug, EnumSetType)]
-pub(crate) enum RxEvent {
-    FifoFull,
-    CmdCharDetected,
-    FifoOvf,
-    FifoTout,
-    GlitchDetected,
-    FrameError,
-    ParityError,
-    BreakDetected,
-}
-
-fn rx_event_check_for_error(events: EnumSet<RxEvent>) -> Result<(), RxError> {
-    for event in events {
-        match event {
-            RxEvent::FifoOvf => return Err(RxError::FifoOverflowed),
-            RxEvent::GlitchDetected => return Err(RxError::GlitchOccurred),
-            RxEvent::FrameError => return Err(RxError::FrameFormatViolated),
-            RxEvent::ParityError => return Err(RxError::ParityMismatch),
-            RxEvent::FifoFull
-            | RxEvent::CmdCharDetected
-            | RxEvent::FifoTout
-            | RxEvent::BreakDetected => continue,
-        }
-    }
-
-    Ok(())
-}
-
-/// A future that resolves when the passed interrupt is triggered,
-/// or has been triggered in the meantime (flag set in INT_RAW).
-/// Upon construction the future enables the passed interrupt and when it
-/// is dropped it disables the interrupt again. The future returns the event
-/// that was initially passed, when it resolves.
-#[must_use = "futures do nothing unless you `.await` or poll them"]
-struct UartRxFuture {
-    events: EnumSet<RxEvent>,
-    uart: &'static Info,
-    state: &'static State,
-    registered: bool,
-}
-
-impl UartRxFuture {
-    fn new(uart: impl Instance, events: impl Into<EnumSet<RxEvent>>) -> Self {
-        Self {
-            events: events.into(),
-            uart: uart.info(),
-            state: uart.state(),
-            registered: false,
-        }
-    }
-}
-
-impl core::future::Future for UartRxFuture {
-    type Output = EnumSet<RxEvent>;
-
-    fn poll(
-        mut self: core::pin::Pin<&mut Self>,
-        cx: &mut core::task::Context<'_>,
-    ) -> core::task::Poll<Self::Output> {
-        let events = self.uart.rx_events().intersection(self.events);
-        if !events.is_empty() {
-            self.uart.clear_rx_events(events);
-            Poll::Ready(events)
-        } else {
-            self.state.rx_waker.register(cx.waker());
-            if !self.registered {
-                self.uart.enable_listen_rx(self.events, true);
-                self.registered = true;
-            }
-            Poll::Pending
-        }
-    }
-}
-
-impl Drop for UartRxFuture {
-    fn drop(&mut self) {
-        // Although the isr disables the interrupt that occurred directly, we need to
-        // disable the other interrupts (= the ones that did not occur), as
-        // soon as this future goes out of scope.
-        self.uart.enable_listen_rx(self.events, false);
-    }
-}
-
-#[must_use = "futures do nothing unless you `.await` or poll them"]
-struct UartTxFuture {
-    events: EnumSet<TxEvent>,
-    uart: &'static Info,
-    state: &'static State,
-    registered: bool,
-}
-
-impl UartTxFuture {
-    fn new(uart: impl Instance, events: impl Into<EnumSet<TxEvent>>) -> Self {
-        Self {
-            events: events.into(),
-            uart: uart.info(),
-            state: uart.state(),
-            registered: false,
-        }
-    }
-}
-
-impl core::future::Future for UartTxFuture {
-    type Output = ();
-
-    fn poll(
-        mut self: core::pin::Pin<&mut Self>,
-        cx: &mut core::task::Context<'_>,
-    ) -> core::task::Poll<Self::Output> {
-        let events = self.uart.tx_events().intersection(self.events);
-        if !events.is_empty() {
-            self.uart.clear_tx_events(events);
-            Poll::Ready(())
-        } else {
-            self.state.tx_waker.register(cx.waker());
-            if !self.registered {
-                self.uart.enable_listen_tx(self.events, true);
-                self.registered = true;
-            }
-            Poll::Pending
-        }
-    }
-}
-
-impl Drop for UartTxFuture {
-    fn drop(&mut self) {
-        // Although the isr disables the interrupt that occurred directly, we need to
-        // disable the other interrupts (= the ones that did not occur), as
-        // soon as this future goes out of scope.
-        self.uart.enable_listen_tx(self.events, false);
-    }
-}
-
-#[instability::unstable]
-impl embedded_io_async_06::Read for Uart<'_, Async> {
-    async fn read(&mut self, buf: &mut [u8]) -> Result<usize, Self::Error> {
-        self.read_async(buf).await.map_err(IoError::Rx)
-    }
-
-    async fn read_exact(
-        &mut self,
-        buf: &mut [u8],
-    ) -> Result<(), embedded_io_06::ReadExactError<Self::Error>> {
-        self.read_exact_async(buf)
-            .await
-            .map_err(|e| embedded_io_06::ReadExactError::Other(IoError::Rx(e)))
-    }
-}
-
-#[instability::unstable]
-impl embedded_io_async_06::Read for UartRx<'_, Async> {
-    async fn read(&mut self, buf: &mut [u8]) -> Result<usize, Self::Error> {
-        self.read_async(buf).await
-    }
-
-    async fn read_exact(
-        &mut self,
-        buf: &mut [u8],
-    ) -> Result<(), embedded_io_06::ReadExactError<Self::Error>> {
-        self.read_exact_async(buf)
-            .await
-            .map_err(embedded_io_06::ReadExactError::Other)
-    }
-}
-
-#[instability::unstable]
-impl embedded_io_async_06::Write for Uart<'_, Async> {
-    async fn write(&mut self, buf: &[u8]) -> Result<usize, Self::Error> {
-        self.write_async(buf).await.map_err(IoError::Tx)
-    }
-
-    async fn flush(&mut self) -> Result<(), Self::Error> {
-        self.flush_async().await.map_err(IoError::Tx)
-    }
-}
-
-#[instability::unstable]
-impl embedded_io_async_06::Write for UartTx<'_, Async> {
-    async fn write(&mut self, buf: &[u8]) -> Result<usize, Self::Error> {
-        self.write_async(buf).await
-    }
-
-    async fn flush(&mut self) -> Result<(), Self::Error> {
-        self.flush_async().await
-    }
-}
-
-#[instability::unstable]
-impl embedded_io_async_07::Read for Uart<'_, Async> {
-    async fn read(&mut self, buf: &mut [u8]) -> Result<usize, Self::Error> {
-        self.read_async(buf).await.map_err(IoError::Rx)
-    }
-
-    async fn read_exact(
-        &mut self,
-        buf: &mut [u8],
-    ) -> Result<(), embedded_io_07::ReadExactError<Self::Error>> {
-        self.read_exact_async(buf)
-            .await
-            .map_err(|e| embedded_io_07::ReadExactError::Other(IoError::Rx(e)))
-    }
-}
-
-#[instability::unstable]
-impl embedded_io_async_07::Read for UartRx<'_, Async> {
-    async fn read(&mut self, buf: &mut [u8]) -> Result<usize, Self::Error> {
-        self.read_async(buf).await
-    }
-
-    async fn read_exact(
-        &mut self,
-        buf: &mut [u8],
-    ) -> Result<(), embedded_io_07::ReadExactError<Self::Error>> {
-        self.read_exact_async(buf)
-            .await
-            .map_err(embedded_io_07::ReadExactError::Other)
-    }
-}
-
-#[instability::unstable]
-impl embedded_io_async_07::Write for Uart<'_, Async> {
-    async fn write(&mut self, buf: &[u8]) -> Result<usize, Self::Error> {
-        self.write_async(buf).await.map_err(IoError::Tx)
-    }
-
-    async fn flush(&mut self) -> Result<(), Self::Error> {
-        self.flush_async().await.map_err(IoError::Tx)
-    }
-}
-
-#[instability::unstable]
-impl embedded_io_async_07::Write for UartTx<'_, Async> {
-    async fn write(&mut self, buf: &[u8]) -> Result<usize, Self::Error> {
-        self.write_async(buf).await
-    }
-
-    async fn flush(&mut self) -> Result<(), Self::Error> {
-        self.flush_async().await
-    }
-}
-
-/// Interrupt handler for all UART instances
-/// Clears and disables interrupts that have occurred and have their enable
-/// bit set. The fact that an interrupt has been disabled is used by the
-/// futures to detect that they should indeed resolve after being woken up
-#[ram]
-pub(super) fn intr_handler(uart: &Info, state: &State) {
-    let interrupts = uart.regs().int_st().read();
-    let interrupt_bits = interrupts.bits(); // = int_raw & int_ena
-    let rx_wake = interrupts.rxfifo_full().bit_is_set()
-        | interrupts.rxfifo_ovf().bit_is_set()
-        | interrupts.rxfifo_tout().bit_is_set()
-        | interrupts.at_cmd_char_det().bit_is_set()
-        | interrupts.glitch_det().bit_is_set()
-        | interrupts.frm_err().bit_is_set()
-        | interrupts.parity_err().bit_is_set()
-        | interrupts.brk_det().bit_is_set();
-    let tx_wake = interrupts.tx_done().bit_is_set() | interrupts.txfifo_empty().bit_is_set();
-
-    uart.regs()
-        .int_ena()
-        .modify(|r, w| unsafe { w.bits(r.bits() & !interrupt_bits) });
-
-    if tx_wake {
-        state.tx_waker.wake();
-    }
-    if rx_wake {
-        state.rx_waker.wake();
     }
 }
 
@@ -2960,875 +2346,3 @@ pub mod lp_uart {
     }
 }
 
-/// A peripheral singleton compatible with the UART driver.
-pub trait Instance: crate::private::Sealed + any::Degrade {
-    #[doc(hidden)]
-    /// Returns the peripheral data and state describing this UART instance.
-    fn parts(&self) -> (&'static Info, &'static State);
-
-    /// Returns the peripheral data describing this UART instance.
-    #[inline(always)]
-    #[doc(hidden)]
-    fn info(&self) -> &'static Info {
-        self.parts().0
-    }
-
-    /// Returns the peripheral state for this UART instance.
-    #[inline(always)]
-    #[doc(hidden)]
-    fn state(&self) -> &'static State {
-        self.parts().1
-    }
-}
-
-/// Peripheral data describing a particular UART instance.
-#[doc(hidden)]
-#[non_exhaustive]
-#[allow(private_interfaces, reason = "Unstable details")]
-pub struct Info {
-    /// Pointer to the register block for this UART instance.
-    ///
-    /// Use [Self::register_block] to access the register block.
-    pub register_block: *const RegisterBlock,
-
-    /// The system peripheral marker.
-    pub peripheral: crate::system::Peripheral,
-
-    /// UART clock group instance.
-    pub clock_instance: clocks::UartInstance,
-
-    /// Interrupt handler for the asynchronous operations of this UART instance.
-    pub async_handler: InterruptHandler,
-
-    /// TX pin
-    pub tx_signal: OutputSignal,
-
-    /// RX pin
-    pub rx_signal: InputSignal,
-
-    /// CTS (Clear to Send) pin
-    pub cts_signal: InputSignal,
-
-    /// RTS (Request to Send) pin
-    pub rts_signal: OutputSignal,
-}
-
-/// Peripheral state for a UART instance.
-#[doc(hidden)]
-#[non_exhaustive]
-pub struct State {
-    /// Waker for the asynchronous RX operations.
-    pub rx_waker: AtomicWaker,
-
-    /// Waker for the asynchronous TX operations.
-    pub tx_waker: AtomicWaker,
-
-    /// Stores whether the RX half is configured for async operation.
-    pub is_rx_async: AtomicBool,
-
-    /// Stores whether the TX half is configured for async operation.
-    pub is_tx_async: AtomicBool,
-}
-
-impl Info {
-    // Currently we don't support merging adjacent FIFO memory, so the max size is
-    // 128 bytes, the max threshold is 127 bytes.
-    const UART_FIFO_SIZE: u16 = property!("uart.ram_size");
-    const RX_FIFO_MAX_THRHD: u16 = Self::UART_FIFO_SIZE - 1;
-    const TX_FIFO_MAX_THRHD: u16 = Self::RX_FIFO_MAX_THRHD;
-
-    /// Returns the register block for this UART instance.
-    pub fn regs(&self) -> &RegisterBlock {
-        unsafe { &*self.register_block }
-    }
-
-    /// Listen for the given interrupts
-    fn enable_listen(&self, interrupts: EnumSet<UartInterrupt>, enable: bool) {
-        let reg_block = self.regs();
-
-        reg_block.int_ena().modify(|_, w| {
-            for interrupt in interrupts {
-                match interrupt {
-                    UartInterrupt::AtCmd => w.at_cmd_char_det().bit(enable),
-                    UartInterrupt::TxDone => w.tx_done().bit(enable),
-                    UartInterrupt::RxBreakDetected => w.brk_det().bit(enable),
-                    UartInterrupt::RxFifoFull => w.rxfifo_full().bit(enable),
-                    UartInterrupt::RxTimeout => w.rxfifo_tout().bit(enable),
-                };
-            }
-            w
-        });
-    }
-
-    fn interrupts(&self) -> EnumSet<UartInterrupt> {
-        let mut res = EnumSet::new();
-        let reg_block = self.regs();
-
-        let ints = reg_block.int_raw().read();
-
-        if ints.at_cmd_char_det().bit_is_set() {
-            res.insert(UartInterrupt::AtCmd);
-        }
-        if ints.tx_done().bit_is_set() {
-            res.insert(UartInterrupt::TxDone);
-        }
-        if ints.brk_det().bit_is_set() {
-            res.insert(UartInterrupt::RxBreakDetected);
-        }
-        if ints.rxfifo_full().bit_is_set() {
-            res.insert(UartInterrupt::RxFifoFull);
-        }
-        if ints.rxfifo_tout().bit_is_set() {
-            res.insert(UartInterrupt::RxTimeout);
-        }
-
-        res
-    }
-
-    fn clear_interrupts(&self, interrupts: EnumSet<UartInterrupt>) {
-        let reg_block = self.regs();
-
-        reg_block.int_clr().write(|w| {
-            for interrupt in interrupts {
-                match interrupt {
-                    UartInterrupt::AtCmd => w.at_cmd_char_det().clear_bit_by_one(),
-                    UartInterrupt::TxDone => w.tx_done().clear_bit_by_one(),
-                    UartInterrupt::RxBreakDetected => w.brk_det().clear_bit_by_one(),
-                    UartInterrupt::RxFifoFull => w.rxfifo_full().clear_bit_by_one(),
-                    UartInterrupt::RxTimeout => w.rxfifo_tout().clear_bit_by_one(),
-                };
-            }
-            w
-        });
-    }
-
-    fn apply_config(&self, config: &Config) -> Result<(), ConfigError> {
-        config.validate()?;
-        self.change_baud(config)?;
-        self.change_data_bits(config.data_bits);
-        self.change_parity(config.parity);
-        self.change_stop_bits(config.stop_bits);
-        self.change_flow_control(config.sw_flow_ctrl, config.hw_flow_ctrl);
-
-        // Avoid glitch interrupts.
-        self.regs().int_clr().write(|w| unsafe { w.bits(u32::MAX) });
-
-        Ok(())
-    }
-
-    fn enable_listen_tx(&self, events: EnumSet<TxEvent>, enable: bool) {
-        self.regs().int_ena().modify(|_, w| {
-            for event in events {
-                match event {
-                    TxEvent::Done => w.tx_done().bit(enable),
-                    TxEvent::FiFoEmpty => w.txfifo_empty().bit(enable),
-                };
-            }
-            w
-        });
-    }
-
-    fn tx_events(&self) -> EnumSet<TxEvent> {
-        let pending_interrupts = self.regs().int_raw().read();
-        let mut active_events = EnumSet::new();
-
-        if pending_interrupts.tx_done().bit_is_set() {
-            active_events |= TxEvent::Done;
-        }
-        if pending_interrupts.txfifo_empty().bit_is_set() {
-            active_events |= TxEvent::FiFoEmpty;
-        }
-
-        active_events
-    }
-
-    fn clear_tx_events(&self, events: impl Into<EnumSet<TxEvent>>) {
-        let events = events.into();
-        self.regs().int_clr().write(|w| {
-            for event in events {
-                match event {
-                    TxEvent::FiFoEmpty => w.txfifo_empty().clear_bit_by_one(),
-                    TxEvent::Done => w.tx_done().clear_bit_by_one(),
-                };
-            }
-            w
-        });
-    }
-
-    fn enable_listen_rx(&self, events: EnumSet<RxEvent>, enable: bool) {
-        self.regs().int_ena().modify(|_, w| {
-            for event in events {
-                match event {
-                    RxEvent::FifoFull => w.rxfifo_full().bit(enable),
-                    RxEvent::BreakDetected => w.brk_det().bit(enable),
-                    RxEvent::CmdCharDetected => w.at_cmd_char_det().bit(enable),
-
-                    RxEvent::FifoOvf => w.rxfifo_ovf().bit(enable),
-                    RxEvent::FifoTout => w.rxfifo_tout().bit(enable),
-                    RxEvent::GlitchDetected => w.glitch_det().bit(enable),
-                    RxEvent::FrameError => w.frm_err().bit(enable),
-                    RxEvent::ParityError => w.parity_err().bit(enable),
-                };
-            }
-            w
-        });
-    }
-
-    fn rx_events(&self) -> EnumSet<RxEvent> {
-        let pending_interrupts = self.regs().int_raw().read();
-        let mut active_events = EnumSet::new();
-
-        if pending_interrupts.rxfifo_full().bit_is_set() {
-            active_events |= RxEvent::FifoFull;
-        }
-        if pending_interrupts.brk_det().bit_is_set() {
-            active_events |= RxEvent::BreakDetected;
-        }
-        if pending_interrupts.at_cmd_char_det().bit_is_set() {
-            active_events |= RxEvent::CmdCharDetected;
-        }
-        if pending_interrupts.rxfifo_ovf().bit_is_set() {
-            active_events |= RxEvent::FifoOvf;
-        }
-        if pending_interrupts.rxfifo_tout().bit_is_set() {
-            active_events |= RxEvent::FifoTout;
-        }
-        if pending_interrupts.glitch_det().bit_is_set() {
-            active_events |= RxEvent::GlitchDetected;
-        }
-        if pending_interrupts.frm_err().bit_is_set() {
-            active_events |= RxEvent::FrameError;
-        }
-        if pending_interrupts.parity_err().bit_is_set() {
-            active_events |= RxEvent::ParityError;
-        }
-
-        active_events
-    }
-
-    fn clear_rx_events(&self, events: impl Into<EnumSet<RxEvent>>) {
-        let events = events.into();
-        self.regs().int_clr().write(|w| {
-            for event in events {
-                match event {
-                    RxEvent::FifoFull => w.rxfifo_full().clear_bit_by_one(),
-                    RxEvent::BreakDetected => w.brk_det().clear_bit_by_one(),
-                    RxEvent::CmdCharDetected => w.at_cmd_char_det().clear_bit_by_one(),
-
-                    RxEvent::FifoOvf => w.rxfifo_ovf().clear_bit_by_one(),
-                    RxEvent::FifoTout => w.rxfifo_tout().clear_bit_by_one(),
-                    RxEvent::GlitchDetected => w.glitch_det().clear_bit_by_one(),
-                    RxEvent::FrameError => w.frm_err().clear_bit_by_one(),
-                    RxEvent::ParityError => w.parity_err().clear_bit_by_one(),
-                };
-            }
-            w
-        });
-    }
-
-    /// Configures the RX-FIFO threshold
-    ///
-    /// ## Errors
-    ///
-    /// [`ConfigError::RxFifoThresholdNotSupported`] if the provided value is zero
-    /// or exceeds [`Info::RX_FIFO_MAX_THRHD`].
-    fn set_rx_fifo_full_threshold(&self, threshold: u16) -> Result<(), ConfigError> {
-        if threshold == 0 || threshold > Self::RX_FIFO_MAX_THRHD {
-            return Err(ConfigError::RxFifoThresholdNotSupported);
-        }
-
-        self.regs()
-            .conf1()
-            .modify(|_, w| unsafe { w.rxfifo_full_thrhd().bits(threshold as _) });
-
-        Ok(())
-    }
-
-    /// Reads the RX-FIFO threshold
-    #[allow(clippy::useless_conversion)]
-    fn rx_fifo_full_threshold(&self) -> u16 {
-        self.regs().conf1().read().rxfifo_full_thrhd().bits().into()
-    }
-
-    /// Configures the TX-FIFO threshold
-    ///
-    /// ## Errors
-    ///
-    /// [`ConfigError::TxFifoThresholdNotSupported`] if the provided value exceeds
-    /// [`Info::TX_FIFO_MAX_THRHD`].
-    fn set_tx_fifo_empty_threshold(&self, threshold: u16) -> Result<(), ConfigError> {
-        if threshold > Self::TX_FIFO_MAX_THRHD {
-            return Err(ConfigError::TxFifoThresholdNotSupported);
-        }
-
-        self.regs()
-            .conf1()
-            .modify(|_, w| unsafe { w.txfifo_empty_thrhd().bits(threshold as _) });
-
-        Ok(())
-    }
-
-    #[procmacros::doc_replace(
-        "rx_timeout_limit" => {
-            cfg(esp32) => "- Symbol size is fixed to 8, do not pass a value > **0x7F**.",
-            _ => "- The value you pass times the symbol size must be <= **0x3FF**.",
-        }
-    )]
-    /// Configures the Receive Timeout detection setting
-    ///
-    /// ## Arguments
-    ///
-    /// `timeout` - the number of symbols ("bytes") to wait for before
-    /// triggering a timeout. Pass None to disable the timeout.
-    ///
-    /// ## Errors
-    ///
-    /// [`ConfigError::TimeoutTooLong`] if the provided value exceeds
-    /// the maximum value for SOC:
-    /// {rx_timeout_limit}
-    fn set_rx_timeout(&self, timeout: Option<u8>, _symbol_len: u8) -> Result<(), ConfigError> {
-        cfg_if::cfg_if! {
-            if #[cfg(esp32)] {
-                const MAX_THRHD: u8 = 0x7F; // 7 bits
-            } else {
-                const MAX_THRHD: u16 = 0x3FF; // 10 bits
-            }
-        }
-
-        let register_block = self.regs();
-
-        if let Some(timeout) = timeout {
-            // the esp32 counts directly in number of symbols (symbol len fixed to 8)
-            #[cfg(esp32)]
-            let timeout_reg = timeout;
-            // all other count in bits, so we need to multiply by the symbol len.
-            #[cfg(not(esp32))]
-            let timeout_reg = timeout as u16 * _symbol_len as u16;
-
-            if timeout_reg > MAX_THRHD {
-                return Err(ConfigError::TimeoutTooLong);
-            }
-
-            cfg_if::cfg_if! {
-                if #[cfg(esp32)] {
-                    let reg_thrhd = register_block.conf1();
-                } else if #[cfg(uart_version = "2")] {
-                    let reg_thrhd = register_block.tout_conf();
-                } else if #[cfg(uart_version = "1")] {
-                    let reg_thrhd = register_block.mem_conf();
-                }
-            }
-            reg_thrhd.modify(|_, w| unsafe { w.rx_tout_thrhd().bits(timeout_reg) });
-        }
-
-        cfg_if::cfg_if! {
-            if #[cfg(uart_version = "2")] {
-                let reg_en = register_block.tout_conf();
-            } else if #[cfg(uart_version = "1")] {
-                let reg_en = register_block.conf1();
-            }
-        }
-        reg_en.modify(|_, w| w.rx_tout_en().bit(timeout.is_some()));
-
-        self.sync_regs();
-
-        Ok(())
-    }
-
-    fn sync_regs(&self) {
-        sync_regs(self.regs());
-    }
-
-    fn change_baud(&self, config: &Config) -> Result<(), ConfigError> {
-        ClockTree::with(|clocks| {
-            let clock = self.clock_instance;
-
-            let source_config = ClockConfig::new(
-                config.clock_source,
-                #[cfg(any(uart_has_sclk_divider, soc_has_pcr, esp32p4))]
-                0,
-            );
-            let clk = clock.function_clock_config_frequency(clocks, source_config);
-
-            // The UART baud rate clock divider is, depending on the device, either a
-            // 20.4 bit, or a 12.4 bit divider.
-            const FRAC_BITS: u32 = const {
-                let largest_divider: u32 =
-                    property!("clock_tree.uart.baud_rate_generator.fractional").1;
-                ::core::assert!((largest_divider + 1).is_power_of_two());
-                largest_divider.count_ones()
-            };
-            const FRAC_MASK: u32 = (1 << FRAC_BITS) - 1;
-
-            // TODO: this block should only prepare the new clock config, and it should
-            // be applied only after validating the resulting baud rate.
-            cfg_if::cfg_if! {
-                if #[cfg(any(uart_has_sclk_divider, soc_has_pcr, esp32p4))] {
-                    const MAX_DIV: u32 = property!("clock_tree.uart.baud_rate_generator.integral").1;
-                    let clk_div = clk.div_ceil(MAX_DIV).div_ceil(config.baudrate);
-                    debug!("SCLK: {} divider: {}", clk, clk_div);
-
-                    let conf = ClockConfig::new(config.clock_source, clk_div - 1);
-                    let divider = (clk << FRAC_BITS) / (config.baudrate * clk_div);
-                } else {
-                    debug!("SCLK: {}", clk);
-                    let conf = ClockConfig::new(config.clock_source);
-                    let divider = (clk << FRAC_BITS) / config.baudrate;
-                }
-            }
-
-            let divider_integer = divider >> FRAC_BITS;
-            let divider_frag = divider & FRAC_MASK;
-            debug!(
-                "UART CLK divider: {} + {}/16",
-                divider_integer, divider_frag
-            );
-
-            clock.configure_function_clock(clocks, conf);
-            clock.configure_baud_rate_generator(
-                clocks,
-                BaudRateConfig::new(divider_frag, divider_integer),
-            );
-
-            self.sync_regs();
-
-            #[cfg(feature = "unstable")]
-            {
-                let deviation_limit = match config.baudrate_tolerance {
-                    BaudrateTolerance::Exact => 1, // Still allow a tiny deviation
-                    BaudrateTolerance::ErrorPercent(percent) => percent as u32,
-                    _ => return Ok(()),
-                };
-
-                let actual_baud = clock.baud_rate_generator_frequency();
-                if actual_baud == 0 {
-                    return Err(ConfigError::BaudrateNotAchievable);
-                }
-
-                let deviation = (config.baudrate.abs_diff(actual_baud) * 100) / actual_baud;
-                debug!(
-                    "Nominal baud: {}, actual: {}, deviation: {}%",
-                    config.baudrate, actual_baud, deviation
-                );
-
-                if deviation > deviation_limit {
-                    return Err(ConfigError::BaudrateNotAchievable);
-                }
-            }
-
-            Ok(())
-        })
-    }
-
-    fn change_data_bits(&self, data_bits: DataBits) {
-        self.regs()
-            .conf0()
-            .modify(|_, w| unsafe { w.bit_num().bits(data_bits as u8) });
-    }
-
-    fn change_parity(&self, parity: Parity) {
-        self.regs().conf0().modify(|_, w| match parity {
-            Parity::None => w.parity_en().clear_bit(),
-            Parity::Even => w.parity_en().set_bit().parity().clear_bit(),
-            Parity::Odd => w.parity_en().set_bit().parity().set_bit(),
-        });
-    }
-
-    fn change_stop_bits(&self, stop_bits: StopBits) {
-        #[cfg(esp32)]
-        {
-            // workaround for hardware issue, when UART stop bit set as 2-bit mode.
-            if stop_bits == StopBits::_2 {
-                self.regs()
-                    .rs485_conf()
-                    .modify(|_, w| w.dl1_en().bit(stop_bits == StopBits::_2));
-
-                self.regs()
-                    .conf0()
-                    .modify(|_, w| unsafe { w.stop_bit_num().bits(1) });
-            }
-        }
-
-        #[cfg(not(esp32))]
-        self.regs()
-            .conf0()
-            .modify(|_, w| unsafe { w.stop_bit_num().bits(stop_bits as u8 + 1) });
-    }
-
-    fn change_flow_control(&self, sw_flow_ctrl: SwFlowControl, hw_flow_ctrl: HwFlowControl) {
-        // set SW flow control
-        match sw_flow_ctrl {
-            SwFlowControl::Enabled {
-                xon_char,
-                xoff_char,
-                xon_threshold,
-                xoff_threshold,
-            } => {
-                cfg_if::cfg_if! {
-                    if #[cfg(uart_version = "2")] {
-                        self.regs().swfc_conf0().modify(|_, w| w.xonoff_del().set_bit().sw_flow_con_en().set_bit());
-                        self.regs().swfc_conf1().modify(|_, w| unsafe { w.xon_threshold().bits(xon_threshold).xoff_threshold().bits(xoff_threshold)});
-                        self.regs().swfc_conf0().modify(|_, w| unsafe { w.xon_char().bits(xon_char).xoff_char().bits(xoff_char) });
-                    } else if #[cfg(esp32)]{
-                        self.regs().flow_conf().modify(|_, w| w.xonoff_del().set_bit().sw_flow_con_en().set_bit());
-                        self.regs().swfc_conf().modify(|_, w| unsafe { w.xon_threshold().bits(xon_threshold).xoff_threshold().bits(xoff_threshold) });
-                        self.regs().swfc_conf().modify(|_, w| unsafe { w.xon_char().bits(xon_char).xoff_char().bits(xoff_char) });
-                    } else if #[cfg(uart_version = "1")] {
-                        self.regs().flow_conf().modify(|_, w| w.xonoff_del().set_bit().sw_flow_con_en().set_bit());
-                        self.regs().swfc_conf1().modify(|_, w| unsafe { w.xon_threshold().bits(xon_threshold as u16) });
-                        self.regs().swfc_conf0().modify(|_, w| unsafe { w.xoff_threshold().bits(xoff_threshold as u16) });
-                        self.regs().swfc_conf1().modify(|_, w| unsafe { w.xon_char().bits(xon_char) });
-                        self.regs().swfc_conf0().modify(|_, w| unsafe { w.xoff_char().bits(xoff_char) });
-                    }
-                }
-            }
-            SwFlowControl::Disabled => {
-                cfg_if::cfg_if! {
-                    if #[cfg(uart_version = "2")] {
-                        let reg = self.regs().swfc_conf0();
-                    } else if #[cfg(uart_version = "1")] {
-                        let reg = self.regs().flow_conf();
-                    }
-                }
-
-                reg.modify(|_, w| w.sw_flow_con_en().clear_bit());
-                reg.modify(|_, w| w.xonoff_del().clear_bit());
-            }
-        }
-
-        self.regs().conf0().modify(|_, w| {
-            w.tx_flow_en()
-                .bit(matches!(hw_flow_ctrl.cts, CtsConfig::Enabled))
-        });
-
-        match hw_flow_ctrl.rts {
-            RtsConfig::Enabled(threshold) => self.configure_rts_flow_ctrl(true, Some(threshold)),
-            RtsConfig::Disabled => self.configure_rts_flow_ctrl(false, None),
-        }
-
-        sync_regs(self.regs());
-    }
-
-    fn configure_rts_flow_ctrl(&self, enable: bool, threshold: Option<u8>) {
-        if let Some(threshold) = threshold {
-            cfg_if::cfg_if! {
-                if #[cfg(esp32)] {
-                    self.regs().conf1().modify(|_, w| unsafe { w.rx_flow_thrhd().bits(threshold) });
-                } else if #[cfg(uart_version = "2")] {
-                    self.regs().hwfc_conf().modify(|_, w| unsafe { w.rx_flow_thrhd().bits(threshold) });
-                } else if #[cfg(uart_version = "1")] {
-                    self.regs().mem_conf().modify(|_, w| unsafe { w.rx_flow_thrhd().bits(threshold as u16) });
-                }
-            }
-        }
-
-        cfg_if::cfg_if! {
-            if #[cfg(uart_version = "2")] {
-                self.regs().hwfc_conf().modify(|_, w| {
-                    w.rx_flow_en().bit(enable)
-                });
-            } else if #[cfg(uart_version = "1")] {
-                self.regs().conf1().modify(|_, w| {
-                    w.rx_flow_en().bit(enable)
-                });
-            }
-        }
-    }
-
-    fn rxfifo_reset(&self) {
-        fn rxfifo_rst(reg_block: &RegisterBlock, enable: bool) {
-            reg_block.conf0().modify(|_, w| w.rxfifo_rst().bit(enable));
-            sync_regs(reg_block);
-        }
-
-        rxfifo_rst(self.regs(), true);
-        rxfifo_rst(self.regs(), false);
-    }
-
-    fn txfifo_reset(&self) {
-        fn txfifo_rst(reg_block: &RegisterBlock, enable: bool) {
-            reg_block.conf0().modify(|_, w| w.txfifo_rst().bit(enable));
-            sync_regs(reg_block);
-        }
-
-        txfifo_rst(self.regs(), true);
-        txfifo_rst(self.regs(), false);
-    }
-
-    fn current_symbol_length(&self) -> u8 {
-        let conf0 = self.regs().conf0().read();
-        let data_bits = conf0.bit_num().bits() + 5; // 5 data bits are encoded as variant 0
-        let parity = conf0.parity_en().bit() as u8;
-        let mut stop_bits = conf0.stop_bit_num().bits();
-
-        match stop_bits {
-            1 => {
-                // workaround for hardware issue, when UART stop bit set as 2-bit mode.
-                #[cfg(esp32)]
-                if self.regs().rs485_conf().read().dl1_en().bit_is_set() {
-                    stop_bits = 2;
-                }
-            }
-            // esp-idf also counts 2 bits for settings 1.5 and 2 stop bits
-            _ => stop_bits = 2,
-        }
-
-        1 + data_bits + parity + stop_bits
-    }
-
-    /// Reads one byte from the RX FIFO.
-    ///
-    /// If the FIFO is empty, the value of the returned byte is not specified.
-    fn read_next_from_fifo(&self) -> u8 {
-        fn access_fifo_register<R>(f: impl Fn() -> R) -> R {
-            // https://docs.espressif.com/projects/esp-chip-errata/en/latest/esp32/03-errata-description/esp32/cpu-subsequent-access-halted-when-get-interrupted.html
-            cfg_if::cfg_if! {
-                if #[cfg(esp32)] {
-                    crate::interrupt::free(f)
-                } else {
-                    f()
-                }
-            }
-        }
-
-        let fifo_reg = self.regs().fifo();
-        cfg_if::cfg_if! {
-            if #[cfg(esp32s2)] {
-                // On the ESP32-S2 we need to use PeriBus2 to read the FIFO:
-                let fifo_reg = unsafe {
-                    &*fifo_reg.as_ptr().cast::<u8>().add(0x20C00000).cast::<crate::pac::uart0::FIFO>()
-                };
-            }
-        }
-
-        access_fifo_register(|| fifo_reg.read().rxfifo_rd_byte().bits())
-    }
-
-    #[allow(clippy::useless_conversion)]
-    fn tx_fifo_count(&self) -> u16 {
-        u16::from(self.regs().status().read().txfifo_cnt().bits())
-    }
-
-    fn write_byte(&self, byte: u8) {
-        self.regs()
-            .fifo()
-            .write(|w| unsafe { w.rxfifo_rd_byte().bits(byte) });
-    }
-
-    fn check_for_errors(&self) -> Result<(), RxError> {
-        let errors = RxEvent::FifoOvf
-            | RxEvent::FifoTout
-            | RxEvent::GlitchDetected
-            | RxEvent::FrameError
-            | RxEvent::ParityError;
-        let events = self.rx_events().intersection(errors);
-        let result = rx_event_check_for_error(events);
-        if result.is_err() {
-            self.clear_rx_events(errors);
-            if events.contains(RxEvent::FifoOvf) {
-                self.rxfifo_reset();
-            }
-        }
-        result
-    }
-
-    #[cfg(not(esp32))]
-    #[allow(clippy::unnecessary_cast)]
-    fn rx_fifo_count(&self) -> u16 {
-        self.regs().status().read().rxfifo_cnt().bits() as u16
-    }
-
-    #[cfg(esp32)]
-    fn rx_fifo_count(&self) -> u16 {
-        let fifo_cnt = self.regs().status().read().rxfifo_cnt().bits();
-
-        // Calculate the real count based on the FIFO read and write offset address:
-        // https://docs.espressif.com/projects/esp-chip-errata/en/latest/esp32/03-errata-description/esp32/uart-fifo-cnt-indicates-data-length-incorrectly.html
-        let status = self.regs().mem_rx_status().read();
-        let rd_addr: u16 = status.mem_rx_rd_addr().bits();
-        let wr_addr: u16 = status.mem_rx_wr_addr().bits();
-
-        if wr_addr > rd_addr {
-            wr_addr - rd_addr
-        } else if wr_addr < rd_addr {
-            (wr_addr + Info::UART_FIFO_SIZE) - rd_addr
-        } else if fifo_cnt > 0 {
-            Info::UART_FIFO_SIZE
-        } else {
-            0
-        }
-    }
-
-    fn write(&self, data: &[u8]) -> Result<usize, TxError> {
-        if data.is_empty() {
-            return Ok(0);
-        }
-
-        while self.tx_fifo_count() >= Info::UART_FIFO_SIZE {}
-
-        let space = (Info::UART_FIFO_SIZE - self.tx_fifo_count()) as usize;
-        let to_write = space.min(data.len());
-        for &byte in &data[..to_write] {
-            self.write_byte(byte);
-        }
-
-        Ok(to_write)
-    }
-
-    fn read(&self, buf: &mut [u8]) -> Result<usize, RxError> {
-        if buf.is_empty() {
-            return Ok(0);
-        }
-
-        while self.rx_fifo_count() == 0 {
-            // Block until we received at least one byte
-            self.check_for_errors()?;
-        }
-
-        self.read_buffered(buf)
-    }
-
-    fn read_buffered(&self, buf: &mut [u8]) -> Result<usize, RxError> {
-        // Get the count first, to avoid accidentally reading a corrupted byte received
-        // after the error check.
-        let to_read = (self.rx_fifo_count() as usize).min(buf.len());
-        self.check_for_errors()?;
-
-        for byte_into in buf[..to_read].iter_mut() {
-            *byte_into = self.read_next_from_fifo();
-        }
-
-        // This bit is not cleared until the FIFO actually drops below the threshold.
-        self.clear_rx_events(RxEvent::FifoFull);
-
-        Ok(to_read)
-    }
-}
-
-impl PartialEq for Info {
-    fn eq(&self, other: &Self) -> bool {
-        core::ptr::eq(self.register_block, other.register_block)
-    }
-}
-
-unsafe impl Sync for Info {}
-
-for_each_uart! {
-    ($id:literal, $inst:ident, $peri:ident, $rxd:ident, $txd:ident, $cts:ident, $rts:ident) => {
-        impl Instance for crate::peripherals::$inst<'_> {
-            fn parts(&self) -> (&'static Info, &'static State) {
-                #[handler]
-                #[ram]
-                pub(super) fn irq_handler() {
-                    intr_handler(&PERIPHERAL, &STATE);
-                }
-
-                static STATE: State = State {
-                    tx_waker: AtomicWaker::new(),
-                    rx_waker: AtomicWaker::new(),
-                    is_rx_async: AtomicBool::new(false),
-                    is_tx_async: AtomicBool::new(false),
-                };
-
-                static PERIPHERAL: Info = Info {
-                    register_block: crate::peripherals::$inst::ptr(),
-                    peripheral: crate::system::Peripheral::$peri,
-                    clock_instance: clocks::UartInstance::$peri,
-                    async_handler: irq_handler,
-                    tx_signal: OutputSignal::$txd,
-                    rx_signal: InputSignal::$rxd,
-                    cts_signal: InputSignal::$cts,
-                    rts_signal: OutputSignal::$rts,
-                };
-                (&PERIPHERAL, &STATE)
-            }
-        }
-    };
-}
-
-crate::any_peripheral! {
-    /// Any UART peripheral.
-    pub peripheral AnyUart<'d> {
-        #[cfg(soc_has_uart0)]
-        Uart0(crate::peripherals::UART0<'d>),
-        #[cfg(soc_has_uart1)]
-        Uart1(crate::peripherals::UART1<'d>),
-        #[cfg(soc_has_uart2)]
-        Uart2(crate::peripherals::UART2<'d>),
-        #[cfg(soc_has_uart3)]
-        Uart3(crate::peripherals::UART3<'d>),
-        #[cfg(soc_has_uart4)]
-        Uart4(crate::peripherals::UART4<'d>),
-    }
-}
-
-impl Instance for AnyUart<'_> {
-    #[inline]
-    fn parts(&self) -> (&'static Info, &'static State) {
-        any::delegate!(self, uart => { uart.parts() })
-    }
-}
-
-impl AnyUart<'_> {
-    fn bind_peri_interrupt(&self, handler: InterruptHandler) {
-        any::delegate!(self, uart => { uart.bind_peri_interrupt(handler) })
-    }
-
-    fn disable_peri_interrupt_on_all_cores(&self) {
-        any::delegate!(self, uart => { uart.disable_peri_interrupt_on_all_cores() })
-    }
-
-    fn set_interrupt_handler(&self, handler: InterruptHandler) {
-        self.disable_peri_interrupt_on_all_cores();
-
-        self.info().enable_listen(EnumSet::all(), false);
-        self.info().clear_interrupts(EnumSet::all());
-
-        self.bind_peri_interrupt(handler);
-    }
-}
-
-struct UartClockGuard<'t> {
-    uart: AnyUart<'t>,
-}
-
-impl<'t> UartClockGuard<'t> {
-    fn new(uart: AnyUart<'t>) -> Self {
-        ClockTree::with(|clocks| {
-            let clock = uart.info().clock_instance;
-
-            // Apply default SCLK configuration
-            let sclk_config = ClockConfig::new(
-                Default::default(),
-                #[cfg(any(uart_has_sclk_divider, soc_has_pcr, esp32p4))]
-                0,
-            );
-            clock.configure_function_clock(clocks, sclk_config);
-            clock.request_function_clock(clocks);
-            clock.request_baud_rate_generator(clocks);
-            #[cfg(soc_has_clock_node_uart_mem_clock)]
-            clock.request_mem_clock(clocks);
-        });
-
-        Self { uart }
-    }
-}
-
-impl Clone for UartClockGuard<'_> {
-    fn clone(&self) -> Self {
-        Self::new(unsafe { self.uart.clone_unchecked() })
-    }
-}
-
-impl Drop for UartClockGuard<'_> {
-    fn drop(&mut self) {
-        ClockTree::with(|clocks| {
-            let clock = self.uart.info().clock_instance;
-
-            #[cfg(soc_has_clock_node_uart_mem_clock)]
-            clock.release_mem_clock(clocks);
-            clock.release_baud_rate_generator(clocks);
-            clock.release_function_clock(clocks);
-        });
-    }
-}
