+++ esp-metadata-generated/src/_generated_esp32.rs
@@ -373,6 +373,9 @@ macro_rules! property {
     ("uart.has_sclk_divider") => {
         false
     };
+    ("uart.has_sclk_enable") => {
+        false
+    };
     ("uhci.combined_uart_selector_field") => {
         false
     };
