+++ esp-metadata-generated/src/_generated_esp32c3.rs
@@ -388,6 +388,9 @@ macro_rules! property {
     ("uart.has_sclk_divider") => {
         true
     };
+    ("uart.has_sclk_enable") => {
+        true
+    };
     ("uhci.combined_uart_selector_field") => {
         false
     };
