+++ esp-metadata-generated/src/_generated_esp32h2.rs
@@ -421,6 +421,9 @@ macro_rules! property {
     ("uart.has_sclk_divider") => {
         false
     };
+    ("uart.has_sclk_enable") => {
+        true
+    };
     ("uhci.combined_uart_selector_field") => {
         false
     };
