+++ esp-metadata-generated/src/_generated_esp32p4.rs
@@ -334,6 +334,9 @@ macro_rules! property {
     ("uart.has_sclk_divider") => {
         false
     };
+    ("uart.has_sclk_enable") => {
+        false
+    };
     ("uhci.combined_uart_selector_field") => {
         false
     };
