+++ esp-metadata-generated/src/_generated_esp32s3.rs
@@ -394,6 +394,9 @@ macro_rules! property {
     ("uart.has_sclk_divider") => {
         true
     };
+    ("uart.has_sclk_enable") => {
+        true
+    };
     ("uhci.combined_uart_selector_field") => {
         false
     };
