+++ esp-metadata/devices/esp32c3.toml
@@ -514,6 +514,7 @@ instances = [
 ram_size = 128
 version = 1
 has_sclk_divider = true
+has_sclk_enable = true
 
 [device.uhci]
 support_status = "partial"
