+++ esp-metadata/devices/esp32c61.toml
@@ -431,6 +431,7 @@ instances = [
 ram_size = 128
 version = 2
 peripheral_controls_mem_clk = true
+has_sclk_enable = true
 
 [device.i2c_master]
 support_status = "supported"
