+++ esp-metadata/devices/esp32s3.toml
@@ -840,6 +840,7 @@ instances = [
 ram_size = 128
 version = 1
 has_sclk_divider = true
+has_sclk_enable = true
 
 [device.uhci]
 support_status = "partial"
